trading-engine/pyproject.toml
[project]
name = "trading-engine"
version = "0.1.0"
description = "Deterministic trading engine for Claude Trade"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "asyncpg>=0.30",
    "sqlmodel>=0.0.22",
    "ib_insync>=0.9.86",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "greenlet>=3.3.2",
    "nest-asyncio>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trading_engine"]
